Feat: Scale UI

+++ core/src/main/java/com/canonal/core/domain/model/UserInfo.kt
@@ -3,7 +3,7 @@ package com.canonal.core.domain.model
 data class UserInfo(
     val gender: Gender,
     val age: Int,
-    val weight: Float,
+    val weight: Int,
     val height: Int,
     val activityLevel: ActivityLevel,
     val goalType: GoalType,
